requirements.txt
asgiref==3.7.2
Django==5.0.2
django-hashid-field==3.4.0
djangorestframework==3.14.0
hashids==1.3.1
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
